pyproject.toml
[tool.poetry]
name = "hierag"
version = "0.1.0"
description = "Core + interfaces monorepo for Hierag"
authors = ["hierag"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = "^0.116.0"
uvicorn = { extras = ["standard"], version = "^0.35.0" }
httpx = "^0.28.1"
fastlite = "*"
beautifulsoup4 = "^4.13.4"
html2text = "^2025.4.15"
lxml = "^6.0.1"
openai = "^2.6.1"
python-dotenv = "^1.1.1"
numpy = "^2.3.3"
sentence-transformers = "^5.1.1"
torch = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
ruff = "^0.12.10"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
